WITCHER: Rename 'script' directory to 'nwscript'
