+++ src/engines/witcher/nwscript/functions.cpp
@@ -44,9 +44,8 @@
 #include "src/engines/witcher/object.h"
 #include "src/engines/witcher/creature.h"
 
-#include "src/engines/witcher/script/functions.h"
-
-#include "src/engines/witcher/script/function_tables.h"
+#include "src/engines/witcher/nwscript/functions.h"
+#include "src/engines/witcher/nwscript/function_tables.h"
 
 namespace Engines {
