+++ src/engines/witcher/nwscript/functions_action.cpp
@@ -38,7 +38,7 @@
 #include "src/engines/witcher/door.h"
 #include "src/engines/witcher/creature.h"
 
-#include "src/engines/witcher/script/functions.h"
+#include "src/engines/witcher/nwscript/functions.h"
 
 namespace Engines {
