+++ src/engines/witcher/nwscript/functions_math.cpp
@@ -27,7 +27,7 @@
 
 #include "src/aurora/nwscript/functioncontext.h"
 
-#include "src/engines/witcher/script/functions.h"
+#include "src/engines/witcher/nwscript/functions.h"
 
 namespace Engines {
