+++ src/engines/witcher/nwscript/functions_module.cpp
@@ -31,7 +31,7 @@
 #include "src/engines/witcher/location.h"
 #include "src/engines/witcher/objectcontainer.h"
 
-#include "src/engines/witcher/script/functions.h"
+#include "src/engines/witcher/nwscript/functions.h"
 
 namespace Engines {
