+++ src/engines/witcher/nwscript/functions_situated.cpp
@@ -28,7 +28,7 @@
 #include "src/engines/witcher/object.h"
 #include "src/engines/witcher/situated.h"
 
-#include "src/engines/witcher/script/functions.h"
+#include "src/engines/witcher/nwscript/functions.h"
 
 namespace Engines {
