+++ src/engines/witcher/nwscript/functions_sound.cpp
@@ -32,7 +32,7 @@
 #include "src/engines/witcher/object.h"
 #include "src/engines/witcher/area.h"
 
-#include "src/engines/witcher/script/functions.h"
+#include "src/engines/witcher/nwscript/functions.h"
 
 namespace Engines {
